Arbi - Fixed many things, added Splash Screen, added ProfileView, updated many common views

+++ willparken/willparken/View/Parkingspot/ParkingspotList.swift
@@ -13,6 +13,10 @@ struct ParkingspotList: View {
     @State private var selectedParkingspotToEdit: Parkingspot?
     @State private var selectedParkingspotToViewReservations: Parkingspot?
     
+    @State private var isLoading: Bool = false
+    @State private var isError: Bool = false
+    @State private var errorMsg: String = ""
+    
     var body: some View {
         List {
             if let parkingspots = wpvm.currentParkingspots{
@@ -22,7 +26,7 @@ struct ParkingspotList: View {
                             Button {
                                 selectedParkingspotToEdit = iParkingspot
                             } label: {
-                                Label("Edit", systemImage: "pencil")
+                                Image(systemName: "pencil")
                             }
                             .tint(.blue)
                         }
@@ -30,12 +34,17 @@ struct ParkingspotList: View {
                             Button {
                                 selectedParkingspotToViewReservations = iParkingspot
                             } label: {
-                                Label("Reservations", systemImage: "calendar.badge.clock")
+                                Image(systemName: "calendar.badge.clock")
                             }
                         }
                 }
                 .onDelete(perform: { indexSet in
                     for index in indexSet {
+                        guard parkingspots[index].p_status != "deleted" else {
+                            errorMsg = "Parkplatz wurde bereits gelöscht."
+                            isError = true
+                            return
+                        }
                         wpvm.deleteParkingspot(parkingspotid: parkingspots[index]._id) { msg in
                             if let msg = msg {
                                 print(msg)
@@ -48,7 +57,7 @@ struct ParkingspotList: View {
             
             Rectangle()
                 .foregroundColor(Color(.black).opacity(0))
-                .frame(height: 50)
+                .frame(height: 80)
                 .listRowSeparator(Visibility.hidden)
         }
         .refreshable {
@@ -61,7 +70,7 @@ struct ParkingspotList: View {
                 .environmentObject(wpvm)
         }
         .sheet(item: $selectedParkingspotToViewReservations, content: { parkingspot in
-            ReservationList(parkingspot: parkingspot)
+            ReservationParkingspotList(parkingspot: parkingspot)
                 .environmentObject(wpvm)
         })
         .sheet(isPresented: $parkingspotCreateOpen){
@@ -74,6 +83,9 @@ struct ParkingspotList: View {
             } label: {
                 Image(systemName: "plus.circle")
             }
+        }
+        .alert(isPresented: $isError) {
+            Alert(title: Text("Oh nein!"), message: Text(errorMsg), dismissButton: .default(Text("OK")))
         }
     }
 }
